+++ alpha/bin/game/character/Character.class
