+++ alpha/bin/game/character/Player.class
