+++ alpha/data/img/characters/player/player_left.png
