+++ alpha/data/img/characters/player/player_left_step_left.png
