+++ alpha/data/img/characters/player/player_right.png
