+++ alpha/data/img/characters/player/player_right_step_right.png
