+++ alpha/data/img/characters/player/player_stat.png
