+++ alpha/data/img/characters/player/player_stat_left.png
