+++ alpha/data/img/characters/player/player_stat_left_step_right.png
