+++ alpha/data/levels/level.tsx
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tileset name="tileset" tilewidth="70" tileheight="70" spacing="2" tilecount="156" columns="12">
+ <image source="../img/tiles_spritesheet.png" width="914" height="936"/>
+ <tile id="2" type="air"/>
+</tileset>
