+++ alpha/src/game/enums/Facing.java
@@ -1,0 +1,7 @@
+package game.enums;
+
+public enum Facing
+{
+	// why use numbers while you can enumerate? :D
+	LEFT, RIGHT
+}
