--- blog project 1/data/img/level.tsx
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<tileset name="tileset" tilewidth="70" tileheight="70" spacing="2" tilecount="156" columns="12">
- <image source="../img/tiles_spritesheet.png" width="914" height="936"/>
-</tileset>
