--- blog project 1/data/levels/empty.tmx
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<map version="1.0" tiledversion="1.0.3" orientation="orthogonal" renderorder="left-up" width="32" height="18" tilewidth="70" tileheight="70" nextobjectid="1">
- <tileset firstgid="1" source="level.tsx"/>
- <layer name="CollisionLayer" width="32" height="18">
-  <data encoding="base64" compression="gzip">
-   H4sIAAAAAAAAC+3DAQkAAAwEoYPv33lBpuCqqao+ek194hwACQAA
-  </data>
- </layer>
-</map>
