--- blog project 1/src/game/enums/Facing.java
@@ -1,7 +1,0 @@
-package game.enums;
-
-public enum Facing
-{
-	// why use numbers while you can enumerate? :D
-	LEFT, RIGHT
-}
